refactor: reorganize GuidedTours directory structure and co-locate CSS
