+++ src/shared/components/GuidedTours/index.ts
@@ -1,8 +1,8 @@
 export { STEP_TYPES, TOUR_ACTIONS, TOUR_STATUS, TOUR_STORAGE_KEY, getTourElement } from './consts';
 export { TourProvider } from './TourProvider';
 export { TourRenderer } from './TourRenderer';
-export { useTour } from './useTour';
-export { useTourAutoTrigger } from './useTourAutoTrigger';
+export { useTour } from './hooks/useTour';
+export { useTourAutoTrigger } from './hooks/useTourAutoTrigger';
 export { registerTour, getToursByRoute } from './registry';
 export type {
   TourConfig,
