+++ src/shared/components/GuidedTours/steps/SpotlightOverlay.scss
@@ -25,11 +25,4 @@
     height: 100vh;
     z-index: var(--pf-t--global--z-index--lg, 499);
   }
-
-  &__highlight-ring {
-    box-shadow: 0 0 0 4px var(--pf-t--global--color--brand--default);
-    border-radius: var(--pf-t--global--border--radius--small);
-    position: relative;
-    z-index: 1;
-  }
 }
